requirements.txt
fastapi==0.65.2
requests==2.25.1
urllib3==1.26.5
uvicorn==0.14.0
httpcore==0.13.4
httpx==0.18.1
python-jose==3.3.0
python-multipart==0.0.5
pydantic==1.8.2
python-dateutil==2.8.1
python-dotenv==0.14.0
python-jose==3.3.0
python-multipart==0.0.5
google-api-python-client==2.15.0
google-cloud-vision==2.4.2